pyproject.toml
[project]
name = "ledgerly-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Leonid Stratulat",email = "leonid.stratulat@cgi.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "reportlab (>=4.4.4,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-keycloak (>=5.8.1,<6.0.0)"
]


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
